requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.0
fastapi-cache2==0.2.1
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
multidict==6.0.5
packaging==23.2
pendulum==3.0.0
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==5.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
time-machine==2.13.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
yarl==1.9.4
